+++ plugins/features/plugin-nprogress/src/node/nprogressPlugin.ts
@@ -1,7 +1,7 @@
 import type { Plugin } from 'vuepress/core'
 import { getDirname, path } from 'vuepress/utils'
 
-const __dirname = getDirname(import.meta.url)
+const __dirname = import.meta.dirname || getDirname(import.meta.url)
 
 export const nprogressPlugin = (): Plugin => ({
   name: '@vuepress/plugin-nprogress',