+++ plugins/markdown/plugin-shiki/src/node/prepareClientConfigFile.ts
@@ -1,9 +1,7 @@
-import { getRealPath } from '@vuepress/helper'
+import { getModulePath } from '@vuepress/helper'
 import type { App } from 'vuepress'
 import type { ShikiPluginOptions } from './options.js'
 import { PLUGIN_NAME } from './utils.js'
-
-const { url } = import.meta
 
 export const prepareClientConfigFile = (
   app: App,
@@ -20,64 +18,64 @@ export const prepareClientConfigFile = (
   }: ShikiPluginOptions,
 ): Promise<string> => {
   const imports: string[] = [
-    `import "${getRealPath('@vuepress/highlighter-helper/styles/base.css', url)}"`,
-    `import "${getRealPath(`${PLUGIN_NAME}/styles/shiki.css`, import.meta.url)}"`,
+    `import "${getModulePath('@vuepress/highlighter-helper/styles/base.css', import.meta)}"`,
+    `import "${getModulePath(`${PLUGIN_NAME}/styles/shiki.css`, import.meta)}"`,
   ]
 
   const setups: string[] = []
 
   if (lineNumbers !== 'disable') {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/line-numbers.css', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/line-numbers.css', import.meta)}"`,
     )
   }
 
   if (highlightLines || notationHighlight) {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/notation-highlight.css', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/notation-highlight.css', import.meta)}"`,
     )
   }
 
   if (notationDiff) {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/notation-diff.css', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/notation-diff.css', import.meta)}"`,
     )
   }
 
   if (notationErrorLevel) {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/notation-error-level.css', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/notation-error-level.css', import.meta)}"`,
     )
   }
 
   if (notationFocus) {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/notation-focus.css', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/notation-focus.css', import.meta)}"`,
     )
   }
 
   if (notationHighlight) {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/notation-highlight.css', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/notation-highlight.css', import.meta)}"`,
     )
   }
 
   if (notationWordHighlight) {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/notation-word-highlight.css', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/notation-word-highlight.css', import.meta)}"`,
     )
   }
 
   if (whitespace) {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/whitespace.css', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/whitespace.css', import.meta)}"`,
     )
   }
 
   if (collapsedLines !== 'disable') {
     imports.push(
-      `import "${getRealPath('@vuepress/highlighter-helper/styles/collapsed-lines.css', url)}"`,
-      `import { setupCollapsedLines } from "${getRealPath('@vuepress/highlighter-helper/composables/collapsedLines.js', url)}"`,
+      `import "${getModulePath('@vuepress/highlighter-helper/styles/collapsed-lines.css', import.meta)}"`,
+      `import { setupCollapsedLines } from "${getModulePath('@vuepress/highlighter-helper/client', import.meta)}"`,
     )
     setups.push('setupCollapsedLines()')
   }