+++ themes/theme-default/src/node/defaultTheme.ts
@@ -24,7 +24,7 @@ import {
   resolveMarkdownHintLocales,
 } from './utils/index.js'
 
-const __dirname = getDirname(import.meta.url)
+const __dirname = import.meta.dirname || getDirname(import.meta.url)
 
 export const defaultTheme = ({
   hostname,