+++ tools/highlighter-helper/rollup.config.ts
@@ -1,9 +1,3 @@
 import { rollupBundle } from '../../scripts/rollup.js'
 
-export default [
-  ...rollupBundle('node/index'),
-  ...rollupBundle({
-    base: 'client',
-    files: ['composables/collapsedLines'],
-  }),
-]
+export default [...rollupBundle('node/index'), ...rollupBundle('client/index')]