+++ CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-rc.64](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.63...v2.0.0-rc.64) (2024-12-07)
+
+### Bug Fixes
+
+- **plugin-docsearch:** fix broken types, close [#296](https://github.com/vuepress/ecosystem/issues/296) ([9b5f904](https://github.com/vuepress/ecosystem/commit/9b5f90452ebe3d179ead3e3fffe28742a168e590))
+- **plugin-theme-data:** fix themeLocaleData type ([dbf0e3d](https://github.com/vuepress/ecosystem/commit/dbf0e3d2292acb411a877b1b04f825d351f0dd6d))
+
+# [2.0.0-rc.63](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.62...v2.0.0-rc.63) (2024-12-06)
+
+**Note:** Version bump only for package @vuepress/ecosystem
+
+# [2.0.0-rc.62](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.61...v2.0.0-rc.62) (2024-12-06)
+
+### Bug Fixes
+
+- **create-vuepress:** prevent navbar overlap in blog template ([#294](https://github.com/vuepress/ecosystem/issues/294)) ([c806fb8](https://github.com/vuepress/ecosystem/commit/c806fb8ecaf582c030a9c8f90de298473495c4a0))
+
+### Features
+
+- **helper:** add sr-only style ([04b856d](https://github.com/vuepress/ecosystem/commit/04b856deffe8beb3e1b6d42e2a760013eafeea55))
+- **plugin-git:** improve commits parsing and optimize plugins ([#292](https://github.com/vuepress/ecosystem/issues/292)) ([a0241c4](https://github.com/vuepress/ecosystem/commit/a0241c4292fc7580c2a5a755d44d7f5df0050e54))
+- **plugin-markdown-hint:** add plugin exports ([8e2aafe](https://github.com/vuepress/ecosystem/commit/8e2aafe6a2fdd8eec6bd16e50e2b8b41bba62d4c))
+
 # [2.0.0-rc.61](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.60...v2.0.0-rc.61) (2024-11-23)
 
 ### Features