+++ plugins/development/plugin-theme-data/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-rc.64](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.63...v2.0.0-rc.64) (2024-12-07)
+
+### Bug Fixes
+
+- **plugin-theme-data:** fix themeLocaleData type ([dbf0e3d](https://github.com/vuepress/ecosystem/commit/dbf0e3d2292acb411a877b1b04f825d351f0dd6d))
+
+# [2.0.0-rc.63](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.62...v2.0.0-rc.63) (2024-12-06)
+
+**Note:** Version bump only for package @vuepress/plugin-theme-data
+
+# [2.0.0-rc.62](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.61...v2.0.0-rc.62) (2024-12-06)
+
+**Note:** Version bump only for package @vuepress/plugin-theme-data
+
 # [2.0.0-rc.61](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.60...v2.0.0-rc.61) (2024-11-23)
 
 **Note:** Version bump only for package @vuepress/plugin-theme-data